test(Geosuggest): re-write google stub to make tests meaningful
Most test cases were broken prior to this fix.
In fact, they were so broken that they seemed to work.
The tests used to set the input value using `component.refs.input.value`. This does not work, and geosuggest will look up '' instead of the updated value.
At the same time, the google stub was incomplete, so that it returned a result for empty strings.
The combination of these two bugs let the tests pass.